Updated openstack/openstack Project: openstack-infra/shade  62616f3eaaff708c583919b55093410a0179c940
Add behavior to enable ironic noauth mode

Added behavior to enable users to utilize the ironic functionality with noauth
mode set on the server side.  This logic is utilized if the auth_plugin is
defined as: None or "None" or ''. Bumped os-client-config requirement
because this working requires the pass-through support that got added
there in 0.5.0.

Change-Id: I3f8a9e0a9952be2d24c6cb655dfed9519134102e