Updated openstack/openstack Project: openstack/cinder  f511194a6d466df25aa0c3ea74f8215bdfd72441
Fixes the import for Lefthand driver

Review 145780 broke the lefthand driver.  The author and reviewers
did not even bother to look at the 3rd party CI results that clearly
showed that this patch broke the hplefthand driver, due to changing
the imports.  The change removed an import to one of the hplefthandclient
modules, which was then later accessed.

Reviewers, please check 3rd party CI when a cinder patch touches a driver.

Authors of patches, if you touch a driver, you need to check the 3rd party
CI failures and make sure your patch changes aren't causing the failures.

Change-Id: Ie38f77fc9498915651c476ea10bc7bd7f49e393d
Partial-Bug: 1407162