Updated openstack/openstack Project: openstack-dev/devstack  172fae6a3ebf867663525048947c797c9209f3eb
Change config of keystone_authtoken in lib/ironic

This commit is temporary work-around until Ironic is fixed(bug#1422632)

The reason of creating temporary work-around is following
Ironic re-uses specific values from the keystone_authtoken middleware.
So we should specify admin_user, admin_password and so on instead of username, password.

Change-Id: If8869e16a167eea0af87afda0eabcbb803627db6
Closes-Bug: #1418341 at devstack