Updated openstack/openstack Project: openstack/trove-specs  bc03edff62663ee0c4e0da1122a3c8a3c0c2ed11
Add support for a new datastore(CouchDB) in Trove

This is to add a new CouchDB plugin for Trove in addition to the
existing NoSQL datastores supported currently. The guestagent for
this will include some of the basic functions like create/delete/
list databases and users as well as create and restore backups.

Change-Id: Ie828f3aac448de0adc1cef1698e7026f38133e6e
Implements: blueprint couchdb-plugin-trove