Updated openstack/openstack Project: openstack/sahara  5597ef29296fbfbb7aaef3eb69acb21d0a2a5547
Applying event log feature for CDH - part 1

It would be nice to apply event log feature for CDH.
Because it's require big changes it would done in
series of small CRs.

This part introduces refactoring, which is required
for applying event log feature. Also, we apply event
log in few places, where it was very easy.

partially implements bp event-log

Change-Id: I4fd2962702d95fef51a8deeda4448dab7294cd87