Updated openstack/openstack Project: openstack-infra/yaml2ical  8bc6cf87c5edd4dd3a6252f056f9074818f6d170
Add ability to generate a meeting index

Adding optional parameters -t and -w used to generate an index of
meetings using a Jinja2 template.

Refactor output file/dir preparation so that index generation can make
use of it.

Expose UTC time in the Schedule object and human-friendly description of
the recurrence in the Recurrence objects, so that they are directly
usable in the templates.

Change-Id: I103ac552f43a02a0b6e2e3ad6e5f52a6877efa68