Updated openstack/openstack Project: openstack/python-manilaclient  717f4245f8541cd7995dcb6b291c5fdd6df0c07c
Make extra spec driver_handles_share_servers required

Currently, the driver_handles_share_servers is an optional extra_spec
on each share_type that allows the scheduler to filter. We want to
make this a required extra spec, so that every share type has this
value set to true or false. It's not be possible to remove that extra
spec. Modifying the extra spec is allowed as long as the value
remains either true or false.

- Add required argument "spec_driver_handles_share_servers" to
"type_create" command
- Add column "required_extra_specs" to columns printed by command
"type-list"
- Change title of column "extra_specs" to "all_extra_specs" in
"extra-specs-list" command
- Add appropriate unit tests

Partially implements bp share-type-require-driver-mode

Change-Id: I9bd3855e8a51159cf5ccc282bcbef688f7268e5e