Updated openstack/openstack Project: openstack-infra/project-config  59ee146ce6ec92b5894648fae54b078a6e6abc60
Use sudo when manipulating git for image list

When we build our images we attempt to checkout all of the valid
devstack branches so that we can list all of the images necessary to be
cached for each branch. Unfortunately this devstack repo is owned by
root during dib and snapshot image builds. This means we must use sudo
when updating the repo (either changing branches or updating the
current branch).

Change-Id: I8cd09cfed4d586648dcbd34fa04bfc030c31ee45