Updated openstack/openstack Project: openstack/barbican  894f9bb2ce4a2a05b0bc63c84a64272cc39af842
Clean up test inheritance

With the addition of the 'mixin' class to the API tests, I noticed
that almost all of the functional tests in test_resources and also the
tests in test_resources_policy ended up inheriting from the
aforementioned class. So this commit is barely to clean things up,
setting up the classes in such a way that the base classes now inherit
from the 'mixin'.

Change-Id: Ibcaff51d2502a8493cf0235a242324f65081b1c5