Updated openstack/openstack Project: openstack-infra/project-config  61a6078839145a50cb1878f3efd2cd7553e9e3e3
Change default behaviour of -nv

Currently the -nv job extension defaults to testing on master only
not stable. There is only one job using the -nv extension at
present. More jobs would like to.

This patch removes the master branch only stipulation on the -nv
extension and places it on the only job using the -nv extension
at present. This enables other jobs to use the -nv extension
to specify job which need to be non-voting. Additional
behaviour for jobs can be specified with different rules.

The -nv jobs default to testing the master branch only:
https://review.openstack.org/#/c/125383/

Change-Id: I76ee5c76cbfc0916538bb2cff7b91996f2a0ff7d