Updated openstack/openstack Project: openstack/nova  2a2c4a3744100f992d9cce40257e282535572f42
Fixes Hyper-V: configdrive is not migrated to destination

When live-migrating an instance with a iso configdrive, that
configdrive is not migrated to the host, even though it is
attached to the instance.

This fix assures that if the instance has a iso configdrive
attached the iso file is also moved during live migration.

Change-Id: Ic8e6b40dcd9a14d5a3fc7f4ef0c5b23c89be599a
Fixes-Bug: 1322096