Updated openstack/openstack Project: openstack/barbican  e2c8ab14da6dc4b8e6b331a6984c622da65fd089
Remove version from endpoints in catalog

Remove the API version from the endpoints in the Keystone service
catalog.  The python-barbicanclient library expects the endpoint to not
include the version, and will add the version itself.  This is
recommended by the Keystone team as a better approach, since the service
catalog does not need to be updated in the event that a new API version
becomes available.

Change-Id: Ibb63113bcbd33d65c691cb242b5794b30114fb23