Updated openstack/openstack Project: openstack/cinder-specs  893c6f862c074e39e745eabe1a67662179a1cbcf
Update of `support-modify-volume-image-metadata.rst`

Since snapshot is immutable, we shouldn't provide the capability
to update image metadata referenced with snapshot.
And part of this Spec is not well formatted, we should update it
to make it easier to read after the rst file is coverted into html
file, overall changes is listed below.

- Remove snapshot relevant changes in the SPEC.
- Fix the incorrect format which will cause the html version
  hard to read.
- Update the work items to reflect the latest changes.

Change-Id: I935ae67bca17c2eaa673d87403c0cfa8068f0922