Updated openstack/openstack Project: openstack/python-heatclient  f213590b81c3b97f65d7137b6675a041419eea41
Implement deployment-create

The deployment-create command creates a derived config
from the supplied config, then creates a deployment with
that. Supported signal transports are TEMP_URL_SIGNAL and
NO_SIGNAL (CFN_SIGNAL and HEAT_SIGNAL are not possible as
they require the deployment to be associated with a stack resource).

The transport TEMP_URL_SIGNAL requires that a swift object
(with TempURL) be created to receive the signal, so heatclient
now depends on swiftclient so that this object can be created.

This change implements the heatclient portion of
blueprint software-config-trigger.

Implements-Blueprint: software-config-trigger

Change-Id: I6d73ba1f0cb9abc3331b5c3d3c788678ad88ff20