Updated openstack/openstack Project: openstack/nova  18a631ca17fcd841d34ecb14d61da2640752e781
Display host chosen for instance by scheduler

The scheduler is configured to schedule an instance on a host chosen
randomly from a subset of the N best hosts. Where the size of the subset
is defined as 'scheduler_host_subset_size' from the configuration.

In a configuration where the subset is greater than the default (1) the
scheduler does not display where it has scheduled the instance to. It
gives a list of all those that it has weighed - just not the one chosen.

This patch displays the chosen host.

Closes-Bug: 1426364

Change-Id: Ibd7d6dcd67609217d5be5e50fbad130676f403d4