Updated openstack/openstack Project: openstack/sahara-specs  fe4815a68c9ce91126ae582b50bdeb803ea7a5df
[EDP] Add Oozie Shell Action job type

Oozie shell actions provide a great deal of flexibility and will
empower users to easily customize and extend the features of Sahara
EDP as needed. For example, a shell action could be used to manage
hdfs on the cluster, do pre or post processing for another job
launched from Sahara, or run a data processing job from a
specialized launcher that does extra configuration not otherwise
available from Sahara (ie, setting a special classpath for a Java job).

blueprint: add-edp-shell-action
Change-Id: Ib975b9ca297611b245c012811b504b2716072c0c