Updated openstack/openstack Project: openstack-infra/release-tools  2f2f552c86834bac4734e72235047c0f5c9c1463
Make git clone work for python-*client projects

Commit 425b6df4f5e5035f0206ffdc601668234fcfc867 added some filtering on
the repo name to account for special cases where the repo name in
git.o.o does not match the project name in launchpad. That prevents
using release_library.sh with projects like python-novaclient.

This adds some logic to check if the project name is of the form
python-*client and if so, assumes the repo name and project in launchpad
are the same. Also updates the comment in the code about the special
case projects for context.

Partial-SB: #2000182

Change-Id: Id459d9e3315bccf1811461eff0095a0f002ae38e