Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  afc25cba892436e1b8242c25afdbef117519247b
Add backup before running db migrations

Previously on new subunit2sql releases we would automatically run the
db upgrade command to cover any potential db migrations added in the
new release. However if there is a potential issue with a new migration
this could potentially corrupt the db. So to offset this risk this
commit runs a mysqldump to backup the db before we run the migration
command so in case things go bad we can restore the db from the backup.

Change-Id: I65d300699c457aeb4ead86c08077f7495eb24ed4