Updated openstack/openstack Project: openstack/cinder  2abbe19ee3874a2e2e0e8fbc7c21d5ee05eb9d5b
Clear migration_status from a destination volume if migration fails

When some error happens during volume migration, cinder-volume
keeps remaining a destination volume because the volume copy may be
in the completing phase and a source volume may be deleted already.

In this case, the migration_status of the destination volume in
data base should be cleared. As a result, the volume can be handled
properly even if volume migration is failed. If the migration is
in the migrating phase, the volume should be deleted and an entry
of database is also destroyed because the source volume is remaining.

Also there is an another problem that a source volume is remaining
as migrating status if create_export raises exception. In this case,
source volume can't be deleted even if it will be unnecessary.

This change fixes these two problems.

Closes-Bug: #1403916
Change-Id: Ie4199bc1e804e5cbbb793f448a99487f9823f1a9