Updated openstack/openstack Project: openstack-infra/storyboard-webclient  a33eaccb4868c4cfd6a0772193b2be3efb6ddc98
Added class and view variable for state transitions.

When we are in the process of transitioning between states, the
application now provides two indicators for UI switching. First,
the class 'resolving' is added to the body element, for CSS targeting.
Secondly, a $rootScope variable named isResolving is set during the
transition, and a root-level loading view has been added.

This permits us to do the lion's share of our data loading in a state's
resolve: {} declaration, and render the final view as a fait accompli.

Change-Id: Ia3580617eba89a0454f383b708c246e4b1901107