Updated openstack/openstack Project: openstack/nova  0ebeb26162f09e3e7dbac08e8ded2097e8628cb1
capture stdout and logging for OSAPIfixture test

the nova fixtures tests do not inherit from nova/test.py because they
are designed to be consumed by it. As such, more manual setup has to
be done per test. Previously in the middle of a test run the
OSAPIFixture was spewing to stdout. This wasn't caught until a new
subunit-trace was released which passed through uncaught stdout.

Change-Id: Ie02403720406cb0ae51a9f5495fb135e1d01992a