Updated openstack/openstack Project: openstack-infra/project-config  ae5a16960a4edcbc80e75fa94fc43a88e74681bb
Limit target projects for the neutron func job

The neutron functional job does not rely on repositories other than
devstack and neutron, so limit the projects that devstack-gate updates
in the workspace accordingly.

This change also reverts the changes to the neutron api job that
validated that the functional job could safely run with a restricted
set of projects.

Change-Id: Ic59c3b2cc5d0b40a8f05c2d89eb1cf5e9e2f0d1d