Updated openstack/openstack Project: openstack/python-barbicanclient  5579cf57ebe4398f2db5a9e784609aadf4901e55
Add secret_type to Secret constructor

The secret_type attribute was recently introduced on the Barbican
server. [1]  Passing a response that includes the secret_type to the
constructor results in a TypeError

This is the kind of error that the upcoming python-barbicanclient gate
will help find.  It should be covered by an existing test in the
functional test suite.

[1] https://review.openstack.org/#/c/156385/

Change-Id: Ibcf42a79ec11e9c05121763ab18d9bc41d6f4dcf
Closes-Bug: #1427423