Updated openstack/openstack Project: openstack/horizon  5f9cb2023eabd297a05040684e2ac179290c4048
Refactor of Metadata Widget

First step of making metadata widget logic reusable.

See dependent patches. It was hard to see the changes
in Gerrit when doing the renaming and refactoring at
once. Gerrit didn't seem to understand that it was
a rename, making the review more difficult to view.

Partial-Bug: 1394472
Change-Id: I01eac21bf78b4b5545405e1ae4cb9a11619e8bc6