Updated openstack/openstack Project: openstack/keystone  cf4ed54a1b18b9b4713ee2e7647dcfe9dc66ad54
Refactor and provide scaffolding for domain specific loading

In preparation for the support of loading domain specific
configuration options from a database backend, the existing
code structure for loading from files is refactored to make
this easier. In addition the config option to support
domain specific configurations loading from a database is
also added (which will result in a NotImplemented error if
it is enabled).

Partially Implements: blueprint domain-config-ext

Change-Id: Iaa7b11e3b719bd68635e7b8a5b7217993db41b1b