Updated openstack/openstack Project: openstack/horizon  025ade43db7038a0f638296fc9c37bd41af8b136
context['view'] not available in Serial Console

In some configurations, the Serial Console gives
a "cannot find instance" error even though the
instance is available for the given id.

See project/instances/views.py SerialConsoleView.
The problem is that it is using context['view'].request
and throwing an exception because in some configurations
context['view'] is not found. Not clear on why it is
there in some cases, e.g. in devstack dev environment.

In any event, the right way to code it is to use
self instead of context['view']. This is what
other classes do in views.py. I confirmed that
self == context['view'].

Closes-bug: #1427391

Change-Id: Ib97f874695c874eb97a5aeb6f84661e33a427477