Updated openstack/openstack Project: openstack/heat  d8b8da59f1583965d215a02fc9ebab9c01a5b510
Utility for database migration.

The DB migration often needs to clone a table into a new one as
databases like sqlite do not support most of the ALTER commands for
in-place modification of table schema. This utility has method to clone
a table to new one so that a new table can be created and data migrated
to it for handling migration. This is fairly generic, therefore new file
is added.

New generic functions can be added here when needed.

Co-Authored-By: Visnusaran Murugan <visnusaran.murugan@hp.com>

Change-Id: I229907dc2cab618ca5c76a72956ce8c5f49a2127