Updated openstack/openstack Project: openstack/taskflow  184b8b49d9ad8a00211f92481e15e5a60d672ba5
Fix lookup scoping multi-match ordering

When look-up is occurring the possible provider
ordering is used instead of the scope returned provider
ordering. This causes incorrect matches when look-up
argument names that are produced by multiple providers (at
the same scope level) all providing the same requirement
name.

The scope order should be enforced as the de-facto order
and not the order that the storage unit finds (which is
hash based and varies depending on hash ordering).

Closes-Bug: #1425326

Change-Id: I15f1ee5515758bdc470c0f7dd7a2f616923e5628