Updated openstack/openstack Project: openstack-dev/devstack  23678b033c2caa3f7b3fd45c47ee79203337649d
Change the Guest Agent log_dir to /var/log/trove

In trove-integration, a patch was submitted to change the value
of the log_dir variable in the trove-guestagent.conf.sample file.
This should be changed here as well so that devstack uses the same
location.

The devstack trove-guestagent.conf file is created by devstack, not
from the sample file, so it needs to be changed here.

(For reference, an actual trove-guestagent.conf file is created by
the cloud provider implementing Trove, who may or may not use the
sample file provided.)

See: https://review.openstack.org/#/c/157973

Change-Id: I41b8e6fe443a2eff376ffc1f92c2ed90c94530b0