Updated openstack/openstack Project: openstack/keystone  ec8f6070abf7576d37837bdf5aec45bc7a055f15
Avoid multiple instances for a provider

Creating multiple instances of a provider leads to unpredictable
results. Instances of dependencies may hold references to any of the
instances.

The dependency module raises an exception if another instance of a
provider is created. The exception has the backtrace where the first
instance was created.

Closes-Bug: #1408384
Change-Id: Idaba61a3cfb851ec7e0e9f7b991f675d77af14e0