Updated openstack/openstack Project: openstack/stevedore  f5eea34ee009686d15f9519426aeead31ab3f78a
Fix test for finding multiple drivers

The mocks in the old version of the test were not bypassing the right
things and the work required to actually bypass them properly was getting
ugly. Rather than mocking, instantiate a valid test instance and then
explicitly call the method that does the validation we're trying to test.

Change-Id: I26a6c6931cc8e95cc00127fd1ad7ab43f396389a