Updated openstack/openstack Project: openstack/openstack-specs  a1c06496a07bf97563bf57d868cdfd528d4c83ce
Add TRACE definition to log guidelines

We should consider actually using the TRACE keyword for logging as
something meaningful. Today TRACE is used as a prefix for stack traces
even though it has no logging level definition.

This would provide a deep debug level which many developers have
requested. TRACE is a standard level in other logging systems such as:
https://logging.apache.org/log4j/1.2/apidocs/org/apache/log4j/Level.html
and means what we are proposing here.

This means that 3rd party tools are highly likely to understand this
meaning for TRACE (and actually be terribly confused by the current
OpenStack use of TRACE).

Co-Authored-By: Alexis Lee <alexisl@hp.com>

Change-Id: Iba69bb07e4c6fd5232efa20b984042a30c85e58b