Updated openstack/openstack Project: openstack/heat  617e8bb3241e79b245acd73d3f32afb1381a8e3e
Remove the template mapping from the env before creating child stacks

Instead of going back through the stack linage, just remove the template
resource "mapping" from the environment when we go to create the
child environment. This is a lot more efficient when creating stacks
with RPC. We don't want to reload every stack in each stack.

part of blueprint decouple-nested
Change-Id: Ic93d08e41d2214e51c836be306a500e5c3939f0f