Updated openstack/openstack Project: openstack/nova  36e8b0ac95b68aab19ce289918db8fc0b106b13d
CWD is incorrectly set if exceptions are thrown

The call to utils.execute ends up in /opt/stack/nova/nova/utils.py which
ultimately calls processutils.execute() in the oslo_concurrency module.
If there's an error when executing the command which calls an bash script
then an exception ProcessExecutionError will be raised.
This patch adds a finally statement to ensure the dir will be switched back
to its original one.

Closes-Bug: #1414530

Change-Id: Ie4e95999795d349a5897f7a180e34187485bd8f1