Updated openstack/openstack Project: openstack/neutron  01530d01dbe8b6b103d0bb72003d566ff51a2fc7
Extend test coverage for iptables_firewall.py

Before refactoring _remove_unused_security_group_info, we
found a few corner cases that were not tested. We ensure
with this test that sg_rules and sg_members of the firewall
object are properly cleaned up when the objects they
keep are not referenced anymore.

Partially implements blueprint refactor-iptables-firewall-driver

Change-Id: I44da9ddddf7ba85ba4b76ecc835155c8fb813d90