Updated openstack/openstack Project: openstack/python-barbicanclient  302e9f36e86b3d94c4c6083721b8aa4101714533
Convert CRUD tests to requests-mock

Use the requests-mock library to stub out the responses that are
returned from the socket rather than mocking out client specific
responses. This makes the tests more flexable as we just ensure that the
correct calls were made to the network and not that certain functions
were called.

Change-Id: I6d9851ef76285de29fc899241cff69f0b8d25af5