Updated openstack/openstack Project: openstack/python-barbicanclient  847baa76f9f4366b06912e7fe525f8a7ab7a8ea1
Remove cyclical dependency

Part of the reason that we typically separate the request manager from
the client object itself as otherwise we end up with a circular
dependency between the client and the managers. This has caused problems
in the past (in other projects) where session objects that maintain a
connection pool were not being freed (and the connections remaining
open) because the circular reference meant that it didn't free the
object.

Create a HTTPClient object that handles the actual requests from
managers.

Change-Id: Ib8a495d6bb7a45e3492aea2e326a7714c79bba8c