Updated openstack/openstack Project: openstack/cinder  2f5eab9ed7548e04869cc4823817480a1cf51f06
Move oslo.messaging to the oslo_messaging namespace

This is the sixth in a series of changes to move to using
the new oslo_<library> namespace that is being used for
oslo libraries.

There is currently a shim in place that is allowing the old
oslo.<library> imports to work, but we need to be prepared for
when the shims go away.  Thus we need patches like this one to
move to the new namespace.

This patch also updates our hacking check to ensure that no
instances of oslo.messaging sneak back in.

Change-Id: Id028f8968717bed49386864081cf150397a6ec5f
Partial-bug: 1409733