Updated openstack/openstack Project: openstack/taskflow  41f64d7cc113f3d33dc3aa394bcec54fd0c49404
Allow node finding to not do a deep search

When a local search is preferred allow for it
to be done by passing in a 'only_direct' option,
also allows passing in a 'include_self' option
to ignore/include the current node when finding.

The defaults of 'include_self' and 'only_direct'
work just as they did prior to this change, so that
the addition does not affect current usage(s) of
this method.

Change-Id: I89fb476bee32144092ea141ca3b971118ab633be